refactor: major style- and component refactor
