+++ src/assets/base.postcss
@@ -1,0 +1,28 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+/* Demo styles -- delete these to start from a clean slate */
+header {
+  @apply mx-auto px-4 pt-6 prose-sm md:px-6 md:prose -mb-8;
+
+  p {
+    @apply text-xl leading-relaxed tracking-wide text-gray-700;
+  }
+}
+
+main {
+  @apply p-4 mx-auto prose md:px-6 prose-indigo sm:rounded-md;
+}
+
+footer {
+  @apply py-6 text-sm text-center text-gray-700;
+
+  a {
+    @apply underline;
+  }
+}
+
+button {
+  @apply px-3 py-2 bg-white border border-gray-300 rounded-md shadow flex items-center;
+}
