+++ src/components/HelloWorld.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
-// useStore, and computed are automatically imported. See vite.config.ts for details.
+// useStore and computed are automatically imported. See vite.config.ts for details.
 const store = useStore()
 const count = computed(() => store.count)
-const props = defineProps<{
+defineProps<{
   msg: string
   optionalProp?: number
 }>()
@@ -10,15 +10,11 @@ const props = defineProps<{
 function increment() {
   store.increment()
 }
-
-defineExpose(props)
 </script>
 
 <template>
-  <h2 class="!mt-0">
-    {{ msg }}
-  </h2>
-  <button class="px-3 py-2 bg-white border border-gray-300 rounded-md shadow flex items-center" @click="increment">
+  <h2>{{ msg }}</h2>
+  <button @click="increment">
     <span class="i-mdi-plus-circle-outline w-5 h-5" />
     <span class="h-7 pl-1">count is: {{ count }}</span>
   </button>
