+++ src/main.ts
@@ -1,9 +1,9 @@
+import '@/assets/base.postcss'
+import router from '@/router'
 import { createHead } from '@unhead/vue'
 import { createPinia } from 'pinia'
 import { createApp, markRaw } from 'vue'
 import App from './App.vue'
-import './assets/index.postcss'
-import router from './router'
 
 const head = createHead()
 const app = createApp(App)
