+++ src/pages/DemoPage.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="p-4 mx-auto prose md:px-6 prose-indigo sm:rounded-md">
+  <div>
     <h2>Demo Page</h2>
 
     <p>
