+++ src/store.ts
@@ -1,12 +1,17 @@
 import { acceptHMRUpdate, defineStore } from 'pinia'
 
-const versionString =
-  import.meta.env.MODE === 'development' ? import.meta.env.VITE_APP_VERSION + '-dev' : import.meta.env.VITE_APP_VERSION
-
 export const useStore = defineStore('main', {
   state: () => ({
     debug: import.meta.env.MODE === 'development',
-    version: versionString,
+    appMeta: {
+      version:
+        import.meta.env.MODE === 'development'
+          ? import.meta.env.VITE_APP_VERSION + '-dev'
+          : import.meta.env.VITE_APP_VERSION,
+      builtAt: import.meta.env.VITE_APP_BUILD_EPOCH
+        ? new Date(Number(import.meta.env.VITE_APP_BUILD_EPOCH))
+        : undefined,
+    },
     isInitialized: false,
     count: 0,
   }),
@@ -29,7 +34,7 @@ export const useStore = defineStore('main', {
 
   getters: {
     isReady: state => {
-      return !state.isInitialized
+      return state.isInitialized
     },
   },
 })
